+++ tests/test_firefish.py
@@ -4,6 +4,7 @@ from unittest import TestCase
 import pytest
 import builtins
 import requests
+import json
 
 
 @pytest.mark.parametrize(
@@ -249,10 +250,27 @@ def test__post_call(
         'poll',
         'quote_id',
         'expected_endpoint',
-        'expected_json'
+        'expected_json',
+        'expected_id'
     ),
     argvalues=[
-        (None, None, None, None, None, None, None, None, None, None, None, None, None, None),
+        (
+            None,
+            None,
+            None,
+            None,
+            None,
+            None,
+            None,
+            None,
+            None,
+            None,
+            None,
+            None,
+            None,
+            None,
+            None
+        ),
         (
             "test content",
             None,
@@ -268,7 +286,8 @@ def test__post_call(
             None,
             "api/notes/create", {
                 "text": "test content"
-            }
+            },
+            123
         ),
         (
             "test content",
@@ -284,7 +303,8 @@ def test__post_call(
             None,
             "api/notes/create", {
                 "text": "test content", "fileIds": [123]
-            }
+            },
+            123
         ),
         (
             "test content",
@@ -301,7 +321,8 @@ def test__post_call(
             None,
             "api/notes/create", {
                 "text": "test content", "visibility": "public"
-            }
+            },
+            123
         ),
         (
             "test content",
@@ -318,7 +339,8 @@ def test__post_call(
             None,
             "api/notes/create", {
                 "text": "test content", "visibility": "followers"
-            }
+            },
+            123
         ),
         (
             "test content",
@@ -335,7 +357,8 @@ def test__post_call(
             None,
             "api/notes/create", {
                 "text": "test content", "visibility": "specified"
-            }
+            },
+            123
         ),
         (
             "test content",
@@ -352,7 +375,8 @@ def test__post_call(
             None,
             "api/notes/create", {
                 "text": "test content", "visibility": "hidden"
-            }
+            },
+            123
         ),
         (
             "test content",
@@ -369,7 +393,8 @@ def test__post_call(
             None,
             "api/notes/create", {
                 "text": "test content", "lang": "ca"
-            }
+            },
+            123
         ),
         (
             "test content",
@@ -386,7 +411,8 @@ def test__post_call(
             None,
             "api/notes/create", {
                 "text": "test content", "replyId": 1234
-            }
+            },
+            123
         ),
     ],
 )
@@ -404,7 +430,8 @@ def test_status_post(
     poll,
     quote_id,
     expected_endpoint,
-    expected_json
+    expected_json,
+    expected_id
 ):
 
     access_token_filename = "user.secret"
@@ -436,8 +463,9 @@ def test_status_post(
             )
     else:
         mocked_post_call = Mock()
+        mocked_post_call.return_value = json.dumps({"id": expected_id})
         with patch.object(instance, "_Firefish__post_call", new=mocked_post_call):
-            _ = instance.status_post(
+            result = instance.status_post(
                 status,
                 in_reply_to_id,
                 media_ids,
@@ -455,3 +483,151 @@ def test_status_post(
             mocked_post_call.assert_called_once_with(
                 endpoint=expected_endpoint, json_data=expected_json
             )
+
+            assert result["id"] == expected_id
+
+
+@pytest.mark.parametrize(
+    argnames=(
+        'media_file',
+        'mime_type',
+        'description',
+        'focus',
+        'file_name',
+        'thumbnail',
+        'thumbnail_mime_type',
+        'synchronous',
+        'expected_endpoint',
+        'expected_json',
+        'expected_id'
+    ),
+    argvalues=[
+        (None, None, None, None, None, None, None, None, None, None, None),
+        (
+            "this/is/my/media.jpg",
+            None,
+            None,
+            None,
+            None,
+            None,
+            None,
+            None,
+            "api/drive/files/create", {
+                "name": "media.jpg"
+            },
+            123
+        ),
+        (
+            "this/is/my/media.jpg",
+            None,
+            "this is alt text",
+            None,
+            None,
+            None,
+            None,
+            None,
+            "api/drive/files/create", {
+                "name": "media.jpg", "comment": "this is alt text"
+            },
+            123
+        ),
+        (
+            b"\x02\x87\x14\xbb\xca\x10\x83\xff\xd9",
+            None,
+            None,
+            None,
+            None,
+            None,
+            None,
+            None,
+            "api/drive/files/create", {},
+            123
+        ),
+        (
+            b"\x02\x87\x14\xbb\xca\x10\x83\xff\xd9",
+            None,
+            None,
+            None,
+            "media.jpg",
+            None,
+            None,
+            None,
+            "api/drive/files/create", {
+                "name": "media.jpg"
+            },
+            123
+        ),
+    ],
+)
+def test_media_post(
+    media_file,
+    mime_type,
+    description,
+    focus,
+    file_name,
+    thumbnail,
+    thumbnail_mime_type,
+    synchronous,
+    expected_endpoint,
+    expected_json,
+    expected_id
+):
+
+    access_token_filename = "user.secret"
+    access_token_client_name = "Client"
+    access_token_api_base_url = "https://social.devnamic.com"
+    access_token_token = "abcdefg123456"
+    access_token_content = "\n".join(
+        [access_token_api_base_url, access_token_client_name, access_token_token]
+    )
+    mocked_content_file = b"\x02\x87\x14\xbb\xca\x10\x83\xff\xd9"
+
+    with patch.object(builtins, "open", mock_open(read_data=access_token_content)):
+        instance = Firefish(access_token=access_token_filename)
+
+    if media_file is None:
+        with TestCase.assertRaises(instance, RuntimeError):
+            instance.media_post(
+                media_file,
+                mime_type,
+                description,
+                focus,
+                file_name,
+                thumbnail,
+                thumbnail_mime_type,
+                synchronous
+            )
+    else:
+        mocked_post_call = Mock()
+        mocked_post_call.return_value = json.dumps({"id": expected_id})
+        with patch.object(instance, "_Firefish__post_call", new=mocked_post_call):
+            if isinstance(media_file, str):
+                with patch.object(builtins, "open", mock_open(read_data=mocked_content_file)):
+                    result = instance.media_post(
+                        media_file,
+                        mime_type,
+                        description,
+                        focus,
+                        file_name,
+                        thumbnail,
+                        thumbnail_mime_type,
+                        synchronous
+                    )
+            else:
+                result = instance.media_post(
+                    media_file,
+                    mime_type,
+                    description,
+                    focus,
+                    file_name,
+                    thumbnail,
+                    thumbnail_mime_type,
+                    synchronous
+                )
+
+            mocked_post_call.assert_called_once_with(
+                endpoint=expected_endpoint,
+                json_data=expected_json,
+                files={"file": mocked_content_file}
+            )
+            assert result["id"] == expected_id